requirements.txt
faiss-cpu
streamlit==1.17.0
google-generativeai==0.1.0
python-dotenv==0.21.0
langchain==0.0.140
langchain_google_genai==0.0.5
langchain_community==0.0.2
langchain_core==0.0.1
gradio==3.0.19
selenium==4.1.0
pillow==9.0.1
sentence-transformers==2.1.0
numpy==1.22.0
